feat: align with wot-typescript-definitions "0.7.0-NEXT-SNAPSHOT.10" * remove fetch ????? * produce accepting TD string only * remove ThingModel
