+++ packages/binding-coap/package.json
@@ -25,7 +25,7 @@
     "typescript-standard": "0.3.30"
   },
   "dependencies": {
-    "wot-typescript-definitions": "0.7.0-NEXT-SNAPSHOT.9",
+    "wot-typescript-definitions": "0.7.0-NEXT-SNAPSHOT.10",
     "@node-wot/core": "0.7.0-SNAPSHOT.1",
     "rxjs": "5.4.3",
     "coap": "0.21.0",
