+++ packages/binding-http/package.json
@@ -29,7 +29,7 @@
     "typescript-standard": "0.3.30"
   },
   "dependencies": {
-    "wot-typescript-definitions": "0.7.0-NEXT-SNAPSHOT.9",
+    "wot-typescript-definitions": "0.7.0-NEXT-SNAPSHOT.10",
     "@node-wot/td-tools": "0.7.0-SNAPSHOT.1",
     "@node-wot/core": "0.7.0-SNAPSHOT.1",
     "basic-auth": "2.0.1",
