+++ packages/cli/package.json
@@ -18,7 +18,7 @@
   },
   "devDependencies": {
     "@types/node": "10.9.4",
-    "wot-typescript-definitions": "0.7.0-NEXT-SNAPSHOT.9",
+    "wot-typescript-definitions": "0.7.0-NEXT-SNAPSHOT.10",
     "ts-node": "3.3.0",
     "typescript": "2.9.2",
     "typescript-standard": "0.3.30"
