+++ packages/core/package.json
@@ -24,7 +24,7 @@
     "typescript-standard": "0.3.30"
   },
   "dependencies": {
-    "wot-typescript-definitions": "0.7.0-NEXT-SNAPSHOT.9",
+    "wot-typescript-definitions": "0.7.0-NEXT-SNAPSHOT.10",
     "@node-wot/td-tools": "0.7.0-SNAPSHOT.1",
     "uuid": "3.3.2",
     "rxjs": "5.4.3",
